requirements.txt
absl-py==1.2.0
aiohttp==3.8.3
aiosignal==1.2.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
entrypoints==0.4
executing==1.0.0
fastjsonschema==2.16.2
fonttools==4.37.4
fqdn==1.5.1
frozenlist==1.3.1
fsspec==2022.8.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.12.0
google-auth-oauthlib==0.4.6
grpcio==1.49.1
idna==3.4
ipykernel==6.15.3
ipython==8.5.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab==3.5.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.19.0
kaleido==0.2.1
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.2
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.8
nbformat==5.7.3
nest-asyncio==1.5.5
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.0
oauthlib==3.2.1
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.6.2
plotly==5.12.0
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyamg==4.2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.2
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.4
pytorch-lightning==1.6.4
pytz==2022.7.1
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.8.1
Send2Trash==1.8.0
sentry-sdk==1.9.10
setproctitle==1.3.2
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.5.0
tenacity==8.1.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
torch==1.12.1
torchinfo==1.7.1
torchmetrics==0.10.0
tornado==6.2
tqdm==4.64.0
traitlets==5.8.1
typing_extensions==4.3.0
uri-template==1.2.0
urllib3==1.26.12
wandb==0.13.3
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
yarl==1.8.1
